requirements.txt
Django==1.10
django-annoying==0.10.1
django-bootstrap3==7.1.0
django-embed-video==1.1.0
django-simple-menu==1.2.0
Pillow==3.1.1
psycopg2==2.6.1
requests==2.9.1
wheel==0.29.0
django-crispy-forms==1.6.0
